+++ controllers/userController.js
@@ -3,6 +3,9 @@ const crypto = require('crypto')
 const bcryptjs = require('bcryptjs');
 const sendMail = require('./sendMail')
 const joi = require('joi')
+const jwt = require('jsonwebtoken')
+
+
 
 const validator =joi.object({
     name: 
@@ -67,7 +70,6 @@ const userController ={
                 country,
                 email,
                 password,
-                role,
                 from
             } = result
             let user = await User.findOne({email})
@@ -79,7 +81,7 @@ const userController ={
                         .toString('hex')
                     if(from === 'form'){ //from form
                         password = bcryptjs.hashSync(password,10);
-                        user = await new User({ name, lastname, photo, country, email, password: [password], role, from: [from], loggedIn, verified, code }).save()
+                        user = await new User({ name, lastname, photo, country, email, password: [password], role:"user", from: [from], loggedIn, verified, code }).save()
                         sendMail(email,code)
                         res.status(201).json({
                             message: "User signed ✔",
@@ -88,7 +90,7 @@ const userController ={
                     } else{ // from socialmedia
                         password = bcryptjs.hashSync(password,10);
                         verified = true,
-                        user = await new User({ name, lastname, photo, country, email, password: [password], role, from: [from], loggedIn, verified, code }).save()
+                        user = await new User({ name, lastname, photo, country, email, password: [password], role:"user", from: [from], loggedIn, verified, code }).save()
                         res.status(201).json({
                             message: "User signed from "+from,
                             success: true,
@@ -126,7 +128,7 @@ const userController ={
             if (user) {
                 user.verified = true
                 await user.save()
-                res.status("200").redirect(301, 'http://localhost:3000/signin')
+                res.status("200").redirect(301, 'http://my-tinerary-juandream.herokuapp.com/signin')
 
             } else {
                 res.status("404").json({
@@ -160,17 +162,18 @@ const userController ={
                                 name: user.name,
                                 email: user.email,
                                 role: user.role,
-                                from: user.from,
                                 photo:user.photo
                             }
 
                             user.loggedIn = true
                             await user.save()
-                            
+                            const token = jwt.sign({id: user._id}, process.env.KEY_JWT, {expiresIn: 60*60*24})
                             res.status(200).json({
                                 success:true,
-                                response: {user: loginUser},
-                                message: "Welcome " + user.name
+                                response: {
+                                    user: loginUser,
+                                    token:token},
+                                message: `Welcome ${user.name}`
                             })
                         }else{ // if password does not match
                             res.status(401).json({
@@ -188,13 +191,15 @@ const userController ={
                             from: user.from,
                             photo:user.photo
                         }
-
+                        const token = jwt.sign({id: user._id}, process.env.KEY_JWT, {expiresIn: 60*60*24})
                         user.loggedIn = true
                         await user.save()
-                        
+                        console.log (user)
                         res.status(200).json({
                             success:true,
-                            response: {user: loginUser},
+                            response: {user: loginUser,
+                                        token: token    
+                            },
                             message: "Welcome " + user.name
                         })
                     }else{ // if password does not match
@@ -247,8 +252,7 @@ const userController ={
     getUser: async (req, res) => {
         const { id } = req.params
         try {
-            let user = await User.findOne({ _id: id })
-                .populate('itineraries', {name:1, city:1})
+            let user = await User.findOne({ _id: id })  
             if (user) {
                 res.status("200").json({
                     message: "Found ✔",
@@ -298,23 +302,50 @@ const userController ={
         }
     },
     modifyUser: async (req, res)=>{
-        const { id } = req.params
-        let putUser = {}
+        // const { id } = req.params
+        const {email} = req.body
+        const {email: uEmail, role:userRole} = req.user
         try {
-            putUser= await User.findOneAndUpdate({_id:id},req.body,{new:true})
-            if (putUser) {
-                res.status("200").json({
-                    message: "User updated.",
-                    response: putUser,
-                    success: true,
-                })
+            console.log(email)
+            if (uEmail.toString() === email || userRole === "admin") {
+                let putUser = await User.findOne({ email: email })
+                if (putUser) {
+                        let {
+                            name,
+                            lastname,
+                            photo,
+                            country,
+                            role
+                        } = req.body;
+                    if (userRole !== "admin") {
+                            putUser = await User.findOneAndUpdate(
+                            { email:email }, {name,lastname,photo,country}, { new: true })
+                                res.status("200").json({
+                                    message: "User updated.",
+                                    response: putUser,
+                                    success: true,
+                            })
+                    } else if(userRole === "admin") {
+                            putUser = await User.findOneAndUpdate(
+                            { email:email }, {name,lastname,photo,country,role}, { new: true })
+                            res.status("200").json({
+                                message: "User updated.",
+                                response: putUser,
+                                success: true,
+                            })
+                    } else{
+                        res.status("404").json({
+                            message: "this User does not exist.",
+                            success: false,
+                        })
+                    }
             } else {
-                res.status("404").json({
-                    message: "this User does not exist.",
+                res.status("401").json({
+                    message: "Unahutorized",
                     success: false,
                 })
             }
-        } catch (error) {
+        }}catch (error) {
             console.log(error)
             res.status("400").json({
                 message: "Error",
@@ -324,20 +355,50 @@ const userController ={
     },
     removeUser: async (req, res) => {
         const { id } = req.params
-        try {
-            await User.findOneAndDelete({ _id: id })
-            res.status("200").json({
-                message: "You deleted an User.",
-                success: true,
-            })
-        } catch (error) {
-            console.log(error)
-            res.status("400").json({
-                message: "Error",
+        if (req.user !== null) {
+            try {
+                await User.findOneAndDelete({ _id: id })
+                res.status("200").json({
+                    message: "You deleted an User.",
+                    success: true,
+                })
+            } catch (error) {
+                console.log(error)
+                res.status("400").json({
+                    message: "Error",
+                    success: false,
+                })
+            }
+        } else {
+            res.status("401").json({
+                message: "Unahutorized",
                 success: false,
             })
         }
+    },
+    verifyToken: (req, res) => {
+        if (req.user !== null){
+            res.status(200).json({
+                success:true,
+                response:{
+                    user: {
+                        id: req.user.userId,
+                        name: req.user.name,
+                        email: req.user.email,
+                        role: req.user.role,
+                        photo:req.user.photo
+                    }
+                },
+                message: 'Welcome ' + req.user.name+'!'
+            })
+        }else {
+            res.json({
+                success:false,
+                message: "Sign in please!"
+            })
+        }
+        }
     }
-}
+
 
 module.exports =userController;
